orderdb duplicated
